Restructured a couple files.
