+++ src/vs14/CoCo_vs14/CoCo_vs14.csproj
@@ -47,29 +47,22 @@
     <WarningLevel>4</WarningLevel>
   </PropertyGroup>
   <ItemGroup>
-    <Compile Include="..\..\vs15\CoCo\AnalyzingService.cs">
-      <Link>AnalyzingService.cs</Link>
-    </Compile>
+    <Compile Include="..\..\vs15\CoCo\Services\AnalyzingService.cs" Link="Services\AnalyzingService.cs" />
     <Compile Include="..\..\vs15\CoCo\Paths.cs" Link="Paths.cs" />
-    <Compile Include="..\..\vs15\CoCo\ResetValuesProvider.cs">
-      <Link>ResetValuesProvider.cs</Link>
-    </Compile>
-    <Compile Include="..\..\vs15\CoCo\ServicesProvider.cs">
-      <Link>ServicesProvider.cs</Link>
-    </Compile>
-    <Compile Include="..\..\vs15\CoCo\VisualBasicClassifierProvider.cs">
-      <Link>VisualBasicClassifierProvider.cs</Link>
-    </Compile>
+    <Compile Include="..\..\vs15\CoCo\Providers\ResetValuesProvider.cs" Link="Providers\ResetValuesProvider.cs" />
+    <Compile Include="..\..\vs15\CoCo\Providers\ServicesProvider.cs" Link="Providers\ServicesProvider.cs" />
+    <Compile Include="..\..\vs15\CoCo\Providers\VisualBasicClassifierProvider.cs" Link="Providers\VisualBasicClassifierProvider.cs" />
+    <Compile Include="..\..\vs15\CoCo\VsPackage.cs" Link="VsPackage.cs" />
     <Compile Include="..\AssemblyInfoCommon.cs" Link="Properties\AssemblyInfoCommon.cs" />
     <Compile Include="..\..\vs15\CoCo\ClassificationManager.cs" Link="ClassificationManager.cs" />
-    <Compile Include="..\..\vs15\CoCo\CSharpClassifierProvider.cs" Link="CSharpClassifierProvider.cs" />
+    <Compile Include="..\..\vs15\CoCo\Providers\CSharpClassifierProvider.cs" Link="Providers\CSharpClassifierProvider.cs" />
     <Compile Include="..\..\vs15\CoCo\Extensions.cs" Link="Extensions.cs" />
-    <Compile Include="..\..\vs15\CoCo\FormattingService.cs" Link="FormattingService.cs" />
-    <Compile Include="..\..\vs15\CoCo\OptionService.cs" Link="OptionService.cs" />
+    <Compile Include="..\..\vs15\CoCo\Services\FormattingService.cs" Link="Services\FormattingService.cs" />
+    <Compile Include="..\..\vs15\CoCo\Services\OptionService.cs" Link="Services\OptionService.cs" />
+    <Compile Include="Guids.cs" />
     <Compile Include="Migrations\MigrationService_2_0_0.cs" />
     <Compile Include="Migrations\MigrationService_2_3_0.cs" />
-    <Compile Include="VsPackage.cs" />
-    <Compile Include="PresetService.cs" />
+    <Compile Include="Services\PresetService.cs" />
     <Compile Include="Properties\AssemblyInfo.cs" />
   </ItemGroup>
   <ItemGroup>
