+++ src/vs14/CoCo_vs14/Guids.cs
@@ -1,0 +1,7 @@
+﻿namespace CoCo
+{
+    internal static class Guids
+    {
+        public const string Package = "076205f5-bacc-4462-878d-67e6d081eba0";
+    }
+}
