+++ src/vs15/CoCo/CoCo.csproj
@@ -49,19 +49,20 @@
     <Compile Include="..\AssemblyInfoCommon.cs">
       <Link>Properties\AssemblyInfoCommon.cs</Link>
     </Compile>
-    <Compile Include="AnalyzingService.cs" />
+    <Compile Include="Guids.cs" />
+    <Compile Include="Services\AnalyzingService.cs" />
     <Compile Include="Migrations\MigrationService_2_3_0.cs" />
     <Compile Include="ClassificationManager.cs" />
     <Compile Include="Extensions.cs" />
-    <Compile Include="CSharpClassifierProvider.cs" />
-    <Compile Include="FormattingService.cs" />
-    <Compile Include="OptionService.cs" />
+    <Compile Include="Providers\CSharpClassifierProvider.cs" />
+    <Compile Include="Services\FormattingService.cs" />
+    <Compile Include="Services\OptionService.cs" />
     <Compile Include="Paths.cs" />
-    <Compile Include="PresetService.cs" />
+    <Compile Include="Services\PresetService.cs" />
     <Compile Include="Properties\AssemblyInfo.cs" />
-    <Compile Include="ResetValuesProvider.cs" />
-    <Compile Include="ServicesProvider.cs" />
-    <Compile Include="VisualBasicClassifierProvider.cs" />
+    <Compile Include="Providers\ResetValuesProvider.cs" />
+    <Compile Include="Providers\ServicesProvider.cs" />
+    <Compile Include="Providers\VisualBasicClassifierProvider.cs" />
     <Compile Include="VsPackage.cs" />
     <Compile Include="Migrations\MigrationService.2_0_0.cs" />
   </ItemGroup>
