+++ src/vs15/CoCo/Guids.cs
@@ -1,0 +1,7 @@
+﻿namespace CoCo
+{
+    internal static class Guids
+    {
+        public const string Package = "b933474d-306e-434f-952d-a820c849ed07";
+    }
+}
