+++ src/vs15/CoCo/Providers/VisualBasicClassifierProvider.cs
@@ -2,12 +2,13 @@
 using System.ComponentModel.Composition;
 using CoCo.Analyser;
 using CoCo.Analyser.VisualBasic;
+using CoCo.Services;
 using CoCo.Utils;
 using Microsoft.VisualStudio.Text;
 using Microsoft.VisualStudio.Text.Classification;
 using Microsoft.VisualStudio.Utilities;
 
-namespace CoCo
+namespace CoCo.Providers
 {
     /// <summary>
     /// Classifier provider which adds <see cref="VisualBasicClassifierProvider"/> to the set of classifiers.
